Move Encoder.kt to package encoding
