+++ escpos4k/src/commonMain/kotlin/cz/multiplatform/escpos4k/core/Command.kt
@@ -17,6 +17,7 @@
 package cz.multiplatform.escpos4k.core
 
 import cz.multiplatform.escpos4k.core.encoding.charset.Charset
+import cz.multiplatform.escpos4k.core.encoding.encode
 
 internal sealed class Command {
   abstract fun bytes(): ByteArray
